# perf improvments for pyxl