The numbers look to good to be true (the 10x pyxl even improves more), but I worked enough for today and will later watch if it's correct.
```
                                68757ba36cff2aa03c:  9b6022e962071f7c09:
           django_template3.py             2.8s (4)             2.8s (4)  +0.2%
                 pyxl_bench.py             2.6s (4)             2.3s (4)  -10.8%
     sqlalchemy_imperative2.py             2.9s (4)             2.9s (4)  -0.5%
                       geomean                 2.7s                 2.6s  -3.8%

```